Moved alert to vuln file, reload page after modals
